backend/requirements.txt
langchain
langchain-community
langchain-core
langchain-nvidia-ai-endpoints
firecrawl-py
wikipedia-api
fastapi
uvicorn
python-dotenv
